refactor(coordinator/api): Phase 4.1 — split provider.go into per-domain files
Pure-move split (deletion-only on the source; no body/comment/signature changed;
locking, Datadog metrics, and handleComplete settle-before-close ordering
preserved by relocation):
  provider.go  1917 -> 300 lines, keeping handleProviderWS + providerReadLoop
  (the WS accept + read-loop dispatcher) and the challenge-interval consts.

New files: provider_challenge, provider_challenge_verify, provider_register,
provider_acme_trust, provider_inference, provider_complete, provider_attestation,
provider_mdm, provider_attestation_handler, provider_trust_status.

go build/vet + go test ./coordinator/api/... (throwaway Postgres) green.
