+++ coordinator/api/provider_attestation.go
@@ -1,0 +1,200 @@
+package api
+
+import (
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/attestation"
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/registry"
+	"github.com/eigeninference/d-inference/coordinator/saferun"
+	"github.com/eigeninference/d-inference/coordinator/store"
+)
+
+// verifyProviderAttestation verifies a provider's Secure Enclave attestation
+// if one was included in the registration message. If the attestation is valid,
+// the provider is marked as attested. If missing or invalid, the provider is
+// accepted in Open Mode only when no binary hash policy is configured.
+func (s *Server) verifyProviderAttestation(providerID string, provider *registry.Provider, regMsg *protocol.RegisterMessage) {
+	policyConfigured, knownBinaryHashes := s.binaryHashPolicySnapshot()
+	if len(regMsg.Attestation) == 0 {
+		if policyConfigured {
+			s.logger.Warn("provider registered without attestation while binary hash policy is configured",
+				"provider_id", providerID,
+			)
+			provider.SetAttestationResult(&attestation.VerificationResult{
+				Valid: false,
+				Error: "attestation missing",
+			})
+			s.registry.MarkUntrusted(providerID)
+			return
+		}
+		s.logger.Info("provider registered without attestation (Open Mode)",
+			"provider_id", providerID,
+		)
+		return
+	}
+
+	result, err := attestation.VerifyJSON(regMsg.Attestation)
+	if err != nil {
+		s.logger.Warn("failed to parse provider attestation",
+			"provider_id", providerID,
+			"error", err,
+		)
+		if policyConfigured {
+			provider.SetAttestationResult(&attestation.VerificationResult{
+				Valid: false,
+				Error: "attestation invalid",
+			})
+			s.registry.MarkUntrusted(providerID)
+		}
+		return
+	}
+
+	provider.SetAttestationResult(&result)
+
+	if !result.Valid {
+		s.logger.Warn("provider attestation invalid",
+			"provider_id", providerID,
+			"error", result.Error,
+		)
+		if policyConfigured {
+			s.registry.MarkUntrusted(providerID)
+		}
+		return
+	}
+
+	// Bind the WebSocket X25519 key used for E2E text encryption to the
+	// attested Secure Enclave identity. If a provider wants to serve private
+	// text, the attestation must carry the same encryption public key.
+	if regMsg.PublicKey != "" {
+		if result.EncryptionPublicKey == "" {
+			s.logger.Warn("attestation missing encryption key for registered public key",
+				"provider_id", providerID,
+			)
+			result.Valid = false
+			result.Error = "attestation missing encryption public key"
+			provider.SetAttestationResult(&result)
+			if policyConfigured {
+				s.registry.MarkUntrusted(providerID)
+			}
+			return
+		}
+		if result.EncryptionPublicKey != regMsg.PublicKey {
+			s.logger.Warn("attestation encryption key does not match register public key",
+				"provider_id", providerID,
+				"attestation_key", result.EncryptionPublicKey,
+				"register_key", regMsg.PublicKey,
+			)
+			result.Valid = false
+			result.Error = "encryption key mismatch"
+			provider.SetAttestationResult(&result)
+			if policyConfigured {
+				s.registry.MarkUntrusted(providerID)
+			}
+			return
+		}
+	}
+
+	// Verify binary hash against known-good hashes. Once a binary hash policy is
+	// configured, omission is a policy violation, not an Open Mode downgrade.
+	if policyConfigured {
+		if result.BinaryHash == "" {
+			s.logger.Warn("provider binary hash missing while known-good policy is configured",
+				"provider_id", providerID,
+			)
+			result.Valid = false
+			result.Error = "binary hash missing"
+			provider.SetAttestationResult(&result)
+			s.registry.MarkUntrusted(providerID)
+			return
+		}
+		binaryHash, err := normalizeSHA256Hex(result.BinaryHash, "binary_hash")
+		if err != nil || !knownBinaryHashes[binaryHash] {
+			s.logger.Warn("provider binary hash not in known-good list",
+				"provider_id", providerID,
+				"binary_hash", result.BinaryHash,
+			)
+			result.Valid = false
+			result.Error = "binary hash not recognized"
+			provider.SetAttestationResult(&result)
+			s.registry.MarkUntrusted(providerID)
+			return
+		}
+		s.logger.Info("provider binary hash verified",
+			"provider_id", providerID,
+			"binary_hash", registry.TruncHash(result.BinaryHash),
+		)
+	}
+
+	provider.SetAttested(true, registry.TrustSelfSigned)
+	s.sendTrustStatus(provider, registry.TrustSelfSigned, "online", "SE attestation verified, awaiting MDM/ACME upgrade")
+
+	// The SE attestation already proves SIP, Secure Boot, and binary hash —
+	// the same checks a challenge re-verifies. Set LastChallengeVerified so
+	// the provider is immediately routable. The 5-minute challenge cycle will
+	// re-verify and add MDM cross-check for defense-in-depth.
+	// Without this, a freshly connected provider waits up to 5 minutes before
+	// it can serve any requests (until first challenge passes).
+	provider.SetLastChallengeVerified(time.Now())
+
+	s.logger.Info("provider attestation verified (self-signed)",
+		"provider_id", providerID,
+		"hardware_model", result.HardwareModel,
+		"chip_name", result.ChipName,
+		"serial_number", result.SerialNumber,
+		"secure_enclave", result.SecureEnclaveAvailable,
+		"sip_enabled", result.SIPEnabled,
+		"secure_boot", result.SecureBootEnabled,
+		"authenticated_root", result.AuthenticatedRootEnabled,
+		"system_volume_hash", result.SystemVolumeHash,
+		"binary_hash", result.BinaryHash,
+		"trust_level", registry.TrustSelfSigned,
+	)
+
+	// Restore persisted state: if this provider was previously known (by serial
+	// number or SE key), restore trust level, reputation, and account linkage.
+	// Fresh attestation verification still runs (above), but stored reputation
+	// is preserved so routing quality is maintained across coordinator restarts.
+	if s.storedProviders != nil {
+		var storedRec *store.ProviderRecord
+		if result.SerialNumber != "" {
+			storedRec = s.storedProviders[result.SerialNumber]
+		}
+		if storedRec == nil && result.PublicKey != "" {
+			storedRec = s.storedProviders["sekey:"+result.PublicKey]
+		}
+		if storedRec != nil {
+			s.registry.RestoreProviderState(provider, storedRec)
+			s.logger.Info("restored persisted provider state",
+				"provider_id", providerID,
+				"stored_serial", storedRec.SerialNumber,
+				"stored_trust", storedRec.TrustLevel,
+			)
+		}
+	}
+
+	// Deduplicate: if another provider connection exists from the same physical
+	// device (same serial number), disconnect it. This prevents multiple
+	// provider processes on the same machine from registering independently
+	// and competing for a single shared vllm-mlx backend.
+	if result.SerialNumber != "" {
+		s.registry.DisconnectDuplicatesBySerial(providerID, result.SerialNumber)
+	}
+
+	// Persist provider state after attestation verification.
+	// This captures the attestation result, serial number, and trust level.
+	s.registry.PersistProvider(provider)
+
+	// MDM verification: independently verify security posture via MicroMDM.
+	// This upgrades trust from self_signed to hardware if MDM confirms
+	// the device is enrolled and SIP/SecureBoot match.
+	if s.mdmClient != nil && result.SerialNumber != "" {
+		saferun.Go(s.logger, "verifyProviderViaMDM", func() {
+			s.verifyProviderViaMDM(providerID, provider, result)
+		})
+	} else if s.mdmClient != nil && result.SerialNumber == "" {
+		s.logger.Warn("provider attestation has no serial number — cannot verify via MDM",
+			"provider_id", providerID,
+		)
+	}
+}
