+++ coordinator/api/provider_challenge.go
@@ -1,0 +1,172 @@
+package api
+
+import (
+	"context"
+	"crypto/rand"
+	"encoding/base64"
+	"encoding/json"
+	"sync"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/registry"
+	"nhooyr.io/websocket"
+)
+
+// pendingChallenge tracks an outstanding challenge sent to a provider.
+type pendingChallenge struct {
+	nonce      string
+	timestamp  string
+	sentAt     time.Time
+	responseCh chan *protocol.AttestationResponseMessage
+}
+
+// challengeTracker manages pending challenges for provider connections.
+type challengeTracker struct {
+	mu      sync.Mutex
+	pending map[string]*pendingChallenge // keyed by nonce
+}
+
+func newChallengeTracker() *challengeTracker {
+	return &challengeTracker{
+		pending: make(map[string]*pendingChallenge),
+	}
+}
+
+func (ct *challengeTracker) add(nonce string, pc *pendingChallenge) {
+	ct.mu.Lock()
+	defer ct.mu.Unlock()
+	ct.pending[nonce] = pc
+}
+
+func (ct *challengeTracker) remove(nonce string) *pendingChallenge {
+	ct.mu.Lock()
+	defer ct.mu.Unlock()
+	pc := ct.pending[nonce]
+	delete(ct.pending, nonce)
+	return pc
+}
+
+// challengeLoop periodically sends attestation challenges to a provider.
+func (s *Server) challengeLoop(ctx context.Context, conn *websocket.Conn, providerID string, provider *registry.Provider, tracker *challengeTracker) {
+	if s.skipChallenge {
+		return
+	}
+
+	interval := s.challengeInterval
+	if interval == 0 {
+		interval = DefaultChallengeInterval
+	}
+
+	// Send initial challenge immediately so the provider is routable
+	// without waiting for the first ticker interval.
+	s.sendChallenge(ctx, conn, providerID, provider, tracker)
+
+	ticker := time.NewTicker(interval)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+			// Stop only for a hard (non-recoverable) untrust. A transiently
+			// untrusted provider (missed-challenge timeouts) keeps being
+			// challenged so a later passing challenge can restore it.
+			if provider.ChallengeShouldStop() {
+				return
+			}
+			s.sendChallenge(ctx, conn, providerID, provider, tracker)
+		}
+	}
+}
+
+// generateNonce creates a random 32-byte nonce and returns it as base64.
+func generateNonce() (string, error) {
+	nonce := make([]byte, 32)
+	if _, err := rand.Read(nonce); err != nil {
+		return "", err
+	}
+	return base64.StdEncoding.EncodeToString(nonce), nil
+}
+
+// sendChallenge sends an attestation challenge to a provider and waits for the response.
+func (s *Server) sendChallenge(ctx context.Context, conn *websocket.Conn, providerID string, provider *registry.Provider, tracker *challengeTracker) {
+	nonce, err := generateNonce()
+	if err != nil {
+		s.logger.Error("failed to generate challenge nonce", "provider_id", providerID, "error", err)
+		return
+	}
+
+	timestamp := time.Now().UTC().Format(time.RFC3339)
+
+	challenge := protocol.AttestationChallengeMessage{
+		Type:      protocol.TypeAttestationChallenge,
+		Nonce:     nonce,
+		Timestamp: timestamp,
+	}
+
+	data, err := json.Marshal(challenge)
+	if err != nil {
+		s.logger.Error("failed to marshal challenge", "provider_id", providerID, "error", err)
+		return
+	}
+
+	pc := &pendingChallenge{
+		nonce:      nonce,
+		timestamp:  timestamp,
+		sentAt:     time.Now(),
+		responseCh: make(chan *protocol.AttestationResponseMessage, 1),
+	}
+	tracker.add(nonce, pc)
+
+	writeCtx, writeCancel := context.WithTimeout(ctx, 5*time.Second)
+	defer writeCancel()
+	if err := conn.Write(writeCtx, websocket.MessageText, data); err != nil {
+		s.logger.Error("failed to send challenge", "provider_id", providerID, "error", err)
+		tracker.remove(nonce)
+		return
+	}
+	s.ddIncr("attestation.challenges_sent", nil)
+
+	s.logger.Debug("sent attestation challenge", "provider_id", providerID, "nonce", nonce[:8]+"...")
+
+	// Wait for response with timeout.
+	timeout := ChallengeResponseTimeout
+	select {
+	case <-ctx.Done():
+		tracker.remove(nonce)
+		return
+	case resp := <-pc.responseCh:
+		tracker.remove(nonce)
+		if resp == nil {
+			// Channel closed without response
+			s.handleChallengeFailure(providerID, "no response")
+			return
+		}
+		s.verifyChallengeResponse(providerID, provider, pc, resp)
+	case <-time.After(timeout):
+		tracker.remove(nonce)
+		s.handleChallengeFailure(providerID, "timeout")
+	}
+}
+
+// handleAttestationResponse processes an attestation response from a provider.
+func (s *Server) handleAttestationResponse(providerID string, provider *registry.Provider, msg *protocol.AttestationResponseMessage, tracker *challengeTracker) {
+	if provider == nil {
+		s.logger.Warn("attestation response from unregistered provider", "provider_id", providerID)
+		return
+	}
+
+	pc := tracker.remove(msg.Nonce)
+	if pc == nil {
+		s.logger.Warn("attestation response for unknown challenge", "provider_id", providerID, "nonce", msg.Nonce[:8]+"...")
+		return
+	}
+
+	// Send response to the waiting goroutine.
+	select {
+	case pc.responseCh <- msg:
+	default:
+	}
+}
