+++ coordinator/api/provider_challenge_verify.go
@@ -1,0 +1,429 @@
+package api
+
+import (
+	"context"
+	"encoding/json"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/attestation"
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/registry"
+	"github.com/eigeninference/d-inference/coordinator/saferun"
+	"nhooyr.io/websocket"
+)
+
+// verifyChallengeResponse verifies a challenge response from a provider.
+// In addition to verifying the nonce and signature, it checks the fresh
+// SIP status reported by the provider. If SIP has been disabled since
+// registration, the provider is marked untrusted immediately.
+func (s *Server) verifyChallengeResponse(providerID string, provider *registry.Provider, pc *pendingChallenge, resp *protocol.AttestationResponseMessage) {
+	// Verify the nonce matches.
+	if resp.Nonce != pc.nonce {
+		s.handleChallengeFailure(providerID, "nonce mismatch")
+		return
+	}
+
+	// Verify the public key matches the registered key.
+	if provider.PublicKey != "" && resp.PublicKey != provider.PublicKey {
+		s.handleChallengeFailure(providerID, "public key mismatch")
+		return
+	}
+
+	// Verify the signature cryptographically using the provider's Secure
+	// Enclave P-256 public key. The provider signs SHA-256(nonce + timestamp)
+	// with its SE key via eigeninference-enclave CLI.
+	if resp.Signature == "" {
+		s.handleChallengeFailure(providerID, "empty signature")
+		return
+	}
+
+	// statusFieldsTrusted gates whether we treat resp.SIPEnabled,
+	// resp.BinaryHash etc. as authoritative. False means the provider
+	// signed only nonce+timestamp (legacy or downgrade), so the status
+	// fields are advisory and we must not act on them as if they were
+	// cryptographically bound.
+	statusFieldsTrusted := false
+
+	// If the provider has an attested SE public key, verify the signature.
+	// Providers without attestation (TrustNone / Open Mode) skip crypto
+	// verification — their trust is already "none".
+	if provider.AttestationResult != nil && provider.AttestationResult.PublicKey != "" {
+		challengeData := pc.nonce + pc.timestamp
+		if err := attestation.VerifyChallengeSignature(
+			provider.AttestationResult.PublicKey,
+			resp.Signature,
+			challengeData,
+		); err != nil {
+			s.logger.Error("challenge signature verification failed",
+				"provider_id", providerID,
+				"error", err,
+			)
+			s.handleChallengeFailure(providerID, "signature verification failed: "+err.Error())
+			return
+		}
+
+		// Now verify the extended status signature if the provider sent
+		// one. Old providers (pre-v0.3.11) won't — log and continue with
+		// status fields untrusted. Mismatch is fatal: it means either
+		// tampering or the provider is signing a different canonical
+		// payload than this code expects.
+		statusInput := attestation.StatusCanonicalInput{
+			Nonce:             pc.nonce,
+			Timestamp:         pc.timestamp,
+			HypervisorActive:  resp.HypervisorActive,
+			RDMADisabled:      resp.RDMADisabled,
+			SIPEnabled:        resp.SIPEnabled,
+			SecureBootEnabled: resp.SecureBootEnabled,
+			BinaryHash:        resp.BinaryHash,
+			ActiveModelHash:   resp.ActiveModelHash,
+			PythonHash:        resp.PythonHash,
+			RuntimeHash:       resp.RuntimeHash,
+			TemplateHashes:    resp.TemplateHashes,
+			ModelHashes:       resp.ModelHashes,
+		}
+		switch err := attestation.VerifyStatusSignature(
+			provider.AttestationResult.PublicKey,
+			resp.StatusSignature,
+			statusInput,
+		); err {
+		case nil:
+			statusFieldsTrusted = true
+		case attestation.ErrStatusSignatureMissing:
+			s.ddIncr("attestation.challenges", []string{"outcome:status_sig_missing"})
+			s.logger.Warn("provider sent no status_signature — status fields are advisory; upgrade provider to bind them",
+				"provider_id", providerID,
+			)
+		default:
+			s.logger.Error("status signature verification failed — possible tampering or canonical mismatch",
+				"provider_id", providerID,
+				"error", err,
+			)
+			s.handleChallengeFailure(providerID, "status signature verification failed: "+err.Error())
+			return
+		}
+	}
+
+	// Status-field enforcement policy (asymmetric, by design):
+	//
+	// The checks below act on resp.SIPEnabled / SecureBootEnabled /
+	// RDMADisabled / BinaryHash / ActiveModelHash regardless of
+	// statusFieldsTrusted. The asymmetry is intentional during the
+	// v0.3.11 rollout window:
+	//
+	//   - Negative reports (SIP=false, hash mismatch, etc.) ALWAYS mark
+	//     the provider untrusted. Acting on a negative is safe even if
+	//     the field is spoofable: the worst case is a compromised
+	//     provider DoS-ing itself, which we want anyway.
+	//
+	//   - Positive reports (SIP=true, hash matches) are accepted but
+	//     can only be fully trusted when statusFieldsTrusted is true.
+	//     A v0.3.10 provider with a compromised process (but intact SE
+	//     key) can echo a valid nonce signature while lying that
+	//     SIPEnabled=true. We accept this risk during rollout.
+	//
+	// TODO(security/v0.3.13+): Once `attestation_challenges_total{
+	// outcome="status_sig_missing"}` is zero across the fleet for a
+	// week, treat ErrStatusSignatureMissing as a hard challenge failure
+	// (target: 2 release cycles after v0.3.11 GA).
+	s.logger.Debug("attestation challenge response verified",
+		"provider_id", providerID,
+		"status_fields_trusted", statusFieldsTrusted,
+	)
+
+	// Verify fresh SIP status. This signal is mandatory for private text:
+	// an omitted value is not evidence of safety, so fail closed.
+	if resp.SIPEnabled == nil {
+		s.handleChallengeFailure(providerID, "SIP status not reported")
+		return
+	}
+	// If the provider reports SIP disabled, they've rebooted since
+	// registration and are no longer trustworthy. SIP cannot be disabled at
+	// runtime — a reboot into Recovery Mode is required.
+	if !*resp.SIPEnabled {
+		s.logger.Error("provider SIP disabled in challenge response — marking untrusted",
+			"provider_id", providerID,
+		)
+		s.registry.MarkUntrusted(providerID)
+		s.handleChallengeFailure(providerID, "SIP disabled")
+		return
+	}
+
+	// Verify fresh Secure Boot status.
+	if resp.SecureBootEnabled != nil && !*resp.SecureBootEnabled {
+		s.logger.Error("provider Secure Boot disabled in challenge response — marking untrusted",
+			"provider_id", providerID,
+		)
+		s.registry.MarkUntrusted(providerID)
+		s.handleChallengeFailure(providerID, "Secure Boot disabled")
+		return
+	}
+
+	// Verify fresh RDMA status. Reporting remains mandatory so routing and
+	// trust policy can distinguish single-node providers from RDMA-aware
+	// cluster runtimes. RDMA enablement is not itself a challenge failure:
+	// Apple Silicon Thunderbolt RDMA is IOMMU-scoped to registered buffers,
+	// so the security boundary is the signed runtime's buffer-registration
+	// discipline, not a hypervisor flag.
+	if resp.RDMADisabled == nil {
+		s.handleChallengeFailure(providerID, "RDMA status not reported — provider must update to v0.2.0+")
+		return
+	}
+	if !*resp.RDMADisabled {
+		s.logger.Info("provider RDMA enabled — accepting under registered-buffer RDMA policy",
+			"provider_id", providerID,
+			"backend", provider.Backend,
+			"hypervisor_active", resp.HypervisorActive,
+		)
+	}
+
+	// Verify fresh binary hash when a known-good policy is configured. A
+	// reported binary hash only counts when the response is signed by the
+	// provider key from a valid registration attestation.
+	policyConfigured, knownBinaryHashes := s.binaryHashPolicySnapshot()
+	if policyConfigured {
+		attestationResult := provider.AttestationResult
+		if attestationResult == nil || !attestationResult.Valid || attestationResult.PublicKey == "" {
+			s.logger.Error("provider cannot prove binary hash without valid attestation",
+				"provider_id", providerID,
+			)
+			s.registry.MarkUntrusted(providerID)
+			s.handleChallengeFailure(providerID, "valid attestation required for binary hash policy")
+			return
+		}
+		if resp.BinaryHash == "" {
+			s.logger.Error("provider omitted binary hash while known-good policy is configured",
+				"provider_id", providerID,
+			)
+			s.registry.MarkUntrusted(providerID)
+			s.handleChallengeFailure(providerID, "binary hash missing")
+			return
+		}
+		attestedBinaryHash, err := normalizeSHA256Hex(attestationResult.BinaryHash, "attested binary_hash")
+		if err != nil {
+			s.logger.Error("provider attestation has no usable binary hash",
+				"provider_id", providerID,
+				"binary_hash", attestationResult.BinaryHash,
+			)
+			s.registry.MarkUntrusted(providerID)
+			s.handleChallengeFailure(providerID, "attested binary hash missing")
+			return
+		}
+		binaryHash, err := normalizeSHA256Hex(resp.BinaryHash, "binary_hash")
+		if err != nil || !knownBinaryHashes[binaryHash] {
+			s.logger.Error("provider binary hash changed — no longer matches known-good list",
+				"provider_id", providerID,
+				"binary_hash", resp.BinaryHash,
+			)
+			s.registry.MarkUntrusted(providerID)
+			s.handleChallengeFailure(providerID, "binary hash mismatch")
+			return
+		}
+		if binaryHash != attestedBinaryHash {
+			s.logger.Error("provider binary hash changed from registration attestation",
+				"provider_id", providerID,
+				"attested_binary_hash", registry.TruncHash(attestedBinaryHash),
+				"challenge_binary_hash", registry.TruncHash(binaryHash),
+			)
+			s.registry.MarkUntrusted(providerID)
+			s.handleChallengeFailure(providerID, "binary hash changed from registration attestation")
+			return
+		}
+	}
+
+	// Verify active model hash if reported and catalog has expected hash.
+	if resp.ActiveModelHash != "" {
+		// Get the current model from the provider's last heartbeat.
+		provider.Mu().Lock()
+		currentModel := provider.CurrentModel
+		provider.Mu().Unlock()
+
+		if currentModel != "" {
+			expectedHash := s.registry.CatalogWeightHash(currentModel)
+			if expectedHash != "" && resp.ActiveModelHash != expectedHash {
+				s.logger.Error("provider active model hash mismatch — possible model swap",
+					"provider_id", providerID,
+					"model", currentModel,
+					"expected", registry.TruncHash(expectedHash),
+					"got", registry.TruncHash(resp.ActiveModelHash),
+				)
+				s.registry.MarkUntrusted(providerID)
+				s.handleChallengeFailure(providerID, "active model weight hash mismatch")
+				return
+			}
+		}
+	}
+
+	// Verify runtime integrity hashes from the signed challenge response.
+	// Swift providers omit Python/vllm hashes, but must still match manifest
+	// entries for external runtime assets such as mlx.metallib.
+	if s.knownRuntimeManifest != nil {
+		runtimeOK, mismatches := s.verifyRuntimeHashesForBackend(
+			provider.Backend, resp.PythonHash, resp.RuntimeHash, resp.TemplateHashes)
+		provider.Mu().Lock()
+		provider.RuntimeVerified = runtimeOK
+		provider.RuntimeManifestChecked = runtimeOK
+		if resp.PythonHash != "" {
+			provider.PythonHash = resp.PythonHash
+		}
+		if resp.RuntimeHash != "" {
+			provider.RuntimeHash = resp.RuntimeHash
+		}
+		if len(resp.TemplateHashes) > 0 {
+			provider.TemplateHashes = registry.CloneStringMap(resp.TemplateHashes)
+		}
+		provider.Mu().Unlock()
+
+		if !runtimeOK {
+			// Log detailed mismatch info for debugging outages.
+			mismatchDetails := make([]string, 0, len(mismatches))
+			for _, m := range mismatches {
+				mismatchDetails = append(mismatchDetails, m.Component+"="+m.Got)
+			}
+			s.logger.Warn("provider runtime integrity mismatch in challenge response — excluding from routing",
+				"provider_id", providerID,
+				"mismatches", len(mismatches),
+				"details", mismatchDetails,
+				"backend", provider.Backend,
+			)
+			// Send status feedback but do NOT fail the challenge or mark untrusted.
+			// The provider remains connected but is excluded from routing until
+			// it reports matching hashes.
+			if provider.Conn != nil {
+				statusMsg := protocol.RuntimeStatusMessage{
+					Type:       protocol.TypeRuntimeStatus,
+					Verified:   false,
+					Mismatches: mismatches,
+				}
+				statusData, err := json.Marshal(statusMsg)
+				if err == nil {
+					writeCtx, writeCancel := context.WithTimeout(context.Background(), 5*time.Second)
+					_ = provider.Conn.Write(writeCtx, websocket.MessageText, statusData)
+					writeCancel()
+				}
+			}
+			return
+		}
+	}
+
+	provider.Mu().Lock()
+	version := provider.Version
+	provider.Mu().Unlock()
+	if s.minProviderVersion != "" && version != "" && semverLess(version, s.minProviderVersion) {
+		s.logger.Warn("provider version below minimum during challenge revalidation — excluded from routing",
+			"provider_id", providerID,
+			"version", version,
+			"min_version", s.minProviderVersion,
+		)
+		s.ddIncr("provider_version_below_minimum", []string{"gate:challenge_revalidation", "version:" + version})
+		provider.Mu().Lock()
+		provider.RuntimeVerified = false
+		provider.RuntimeManifestChecked = false
+		provider.Mu().Unlock()
+		return
+	}
+
+	// Override self-reported privacy capabilities with coordinator-verified
+	// values from the challenge response. The coordinator independently checks
+	// SIP during each attestation challenge. Hypervisor status is preserved as
+	// a reported capability only; it is not the RDMA safety proof.
+	provider.Mu().Lock()
+	if provider.PrivacyCapabilities != nil {
+		if resp.SIPEnabled != nil {
+			provider.PrivacyCapabilities.SIPEnabled = *resp.SIPEnabled
+		}
+		if resp.HypervisorActive != nil {
+			provider.PrivacyCapabilities.HypervisorActive = *resp.HypervisorActive
+		}
+	}
+	provider.ChallengeVerifiedSIP = resp.SIPEnabled != nil && *resp.SIPEnabled
+	provider.Mu().Unlock()
+
+	// Challenge passed.
+	recovered := s.registry.RecordChallengeSuccess(providerID)
+	if recovered {
+		// The provider was transiently untrusted and is now back online. It was
+		// last told "untrusted" (handleChallengeFailure) and scheduled a 10-min
+		// diagnostic auto-report; push a fresh "online" trust_status so it clears
+		// that local state and cancels the report.
+		provider.Mu().Lock()
+		trustLevel := provider.TrustLevel
+		provider.Mu().Unlock()
+		s.sendTrustStatus(provider, trustLevel, "online", "recovered after transient deroute")
+	}
+	s.ddIncr("attestation.challenges", []string{"outcome:passed"})
+	s.logger.Info("attestation challenge verified",
+		"provider_id", providerID,
+		"sip_enabled", resp.SIPEnabled,
+		"secure_boot_enabled", resp.SecureBootEnabled,
+		"rdma_disabled", resp.RDMADisabled,
+		"hypervisor_active", resp.HypervisorActive,
+		"binary_hash", resp.BinaryHash,
+		"active_model_hash", resp.ActiveModelHash,
+		"model_hashes_count", len(resp.ModelHashes),
+	)
+	for modelID, hash := range resp.ModelHashes {
+		s.logger.Info("model weight hash verified",
+			"provider_id", providerID,
+			"model_id", modelID,
+			"weight_hash", hash,
+		)
+	}
+
+	// Re-attempt MDM verification for self_signed providers. This handles
+	// providers that installed the MDM enrollment profile after their initial
+	// registration — they would otherwise stay at self_signed trust forever
+	// since verifyProviderViaMDM only ran once at registration time.
+	provider.Mu().Lock()
+	trustLevel := provider.TrustLevel
+	attestResult := provider.AttestationResult
+	provider.Mu().Unlock()
+
+	if trustLevel == registry.TrustSelfSigned && s.mdmClient != nil && attestResult != nil {
+		result := *attestResult
+		saferun.Go(s.logger, "retryMDMVerification", func() {
+			s.verifyProviderViaMDM(providerID, provider, result)
+		})
+	}
+}
+
+// handleChallengeFailure records a failed challenge and marks the provider
+// as untrusted if the failure threshold is reached.
+func (s *Server) handleChallengeFailure(providerID string, reason string) {
+	transient := reason == "timeout" || reason == "no response"
+	failures := s.registry.RecordChallengeFailure(providerID, transient)
+	s.ddIncr("attestation.challenges", []string{"outcome:failed"})
+	s.logger.Warn("attestation challenge failed",
+		"provider_id", providerID,
+		"reason", reason,
+		"consecutive_failures", failures,
+	)
+
+	severity := protocol.SeverityWarn
+	if failures >= registry.MaxFailedChallenges {
+		severity = protocol.SeverityError
+		if transient {
+			// Missed-challenge timeouts (sleep / network blip) are recoverable:
+			// keep challenging and let a later passing challenge restore the
+			// provider without requiring a reconnect.
+			s.registry.MarkUntrustedTransient(providerID)
+		} else {
+			s.registry.MarkUntrusted(providerID)
+		}
+		if p := s.registry.GetProvider(providerID); p != nil {
+			s.sendTrustStatus(p, p.TrustLevel, string(registry.StatusUntrusted), reason)
+		}
+	}
+	s.emit(context.Background(), severity, protocol.KindAttestationFailure,
+		"attestation challenge failed",
+		map[string]any{
+			"provider_id":     providerID,
+			"reason":          reason,
+			"reconnect_count": failures,
+		})
+	if s.metrics != nil {
+		s.metrics.IncCounter("attestation_failures_total",
+			MetricLabel{"reason", reason},
+		)
+	}
+	s.ddIncr("attestation.failures", []string{"reason:" + reason})
+}
