+++ coordinator/api/provider_inference.go
@@ -1,0 +1,142 @@
+package api
+
+import (
+	"net/http"
+	"strings"
+
+	"github.com/eigeninference/d-inference/coordinator/internal/e2e"
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/registry"
+)
+
+func (s *Server) handleChunk(providerID string, provider *registry.Provider, msg *protocol.InferenceResponseChunkMessage) {
+	if provider == nil {
+		s.logger.Warn("chunk from unregistered provider", "provider_id", providerID)
+		return
+	}
+	pr := provider.GetPending(msg.RequestID)
+	if pr == nil {
+		s.logger.Warn("chunk for unknown request", "provider_id", providerID, "request_id", msg.RequestID)
+		return
+	}
+	chunkData, err := decryptTextResponseChunk(provider, pr, msg)
+	if err != nil {
+		s.logger.Warn("rejecting insecure response chunk",
+			"provider_id", providerID,
+			"request_id", msg.RequestID,
+			"error", err,
+		)
+		s.registry.MarkUntrusted(providerID)
+		s.handleInferenceError(providerID, provider, &protocol.InferenceErrorMessage{
+			Type:       protocol.TypeInferenceError,
+			RequestID:  msg.RequestID,
+			Error:      "provider returned invalid encrypted chunk",
+			StatusCode: http.StatusBadGateway,
+		})
+		return
+	}
+	// Non-blocking send — if consumer is gone the chunk is dropped.
+	select {
+	case pr.ChunkCh <- chunkData:
+	default:
+		s.logger.Warn("dropped chunk, consumer channel full", "request_id", msg.RequestID)
+	}
+}
+
+func decryptTextResponseChunk(provider *registry.Provider, pr *registry.PendingRequest, msg *protocol.InferenceResponseChunkMessage) (string, error) {
+	if msg.EncryptedData == nil {
+		return "", errTextChunkViolation("plaintext text chunk")
+	}
+	if msg.Data != "" {
+		return "", errTextChunkViolation("mixed plaintext and encrypted text chunk")
+	}
+	if provider.PublicKey == "" {
+		return "", errTextChunkViolation("provider missing registered public key")
+	}
+	if msg.EncryptedData.EphemeralPublicKey != provider.PublicKey {
+		return "", errTextChunkViolation("chunk sender key mismatch")
+	}
+	if pr.SessionPrivKey == nil {
+		return "", errTextChunkViolation("missing coordinator session key")
+	}
+
+	payload := &e2e.EncryptedPayload{
+		EphemeralPublicKey: msg.EncryptedData.EphemeralPublicKey,
+		Ciphertext:         msg.EncryptedData.Ciphertext,
+	}
+	session := &e2e.SessionKeys{PrivateKey: *pr.SessionPrivKey}
+	plaintext, err := e2e.Decrypt(payload, session)
+	if err != nil {
+		return "", err
+	}
+	return string(plaintext), nil
+}
+
+func errTextChunkViolation(reason string) error {
+	return &textChunkViolationError{reason: reason}
+}
+
+type textChunkViolationError struct {
+	reason string
+}
+
+func (e *textChunkViolationError) Error() string {
+	return e.reason
+}
+
+func (s *Server) handleInferenceAccepted(provider *registry.Provider, msg *protocol.InferenceAcceptedMessage) {
+	if provider == nil {
+		return
+	}
+	pr := provider.GetPending(msg.RequestID)
+	if pr == nil {
+		return
+	}
+	// Non-blocking signal — the dispatch loop may have already committed.
+	select {
+	case pr.AcceptedCh <- struct{}{}:
+	default:
+	}
+}
+
+func (s *Server) handleInferenceError(providerID string, provider *registry.Provider, msg *protocol.InferenceErrorMessage) {
+	if provider == nil {
+		s.logger.Warn("error from unregistered provider", "provider_id", providerID)
+		return
+	}
+	pr := provider.RemovePending(msg.RequestID)
+	if pr == nil {
+		s.logger.Warn("error for unknown request", "provider_id", providerID, "request_id", msg.RequestID)
+		return
+	}
+
+	pr.ErrorCh <- *msg
+	close(pr.ChunkCh)
+	close(pr.CompleteCh)
+	close(pr.ErrorCh)
+
+	// Record job failure for reputation tracking, but carve out capacity
+	// rejections — those are not provider faults, just the provider declining
+	// work it cannot currently serve (the coordinator reroutes these). Counting
+	// them would unfairly penalise healthy providers shedding load. Capacity
+	// signals: HTTP 503 (service unavailable) / 429 (too many requests), an
+	// exhausted token budget, or an out-of-memory model-load reject.
+	loweredErr := strings.ToLower(msg.Error)
+	capacityRejection := msg.StatusCode == http.StatusServiceUnavailable ||
+		msg.StatusCode == http.StatusTooManyRequests ||
+		strings.Contains(loweredErr, "token_budget_exhausted") ||
+		strings.Contains(loweredErr, "insufficient memory")
+	if !capacityRejection {
+		s.registry.RecordJobFailure(providerID)
+	}
+
+	// Mark provider idle.
+	s.registry.SetProviderIdle(providerID)
+
+	s.logger.Error("inference error",
+		"request_id", msg.RequestID,
+		"provider_id", providerID,
+		"error", msg.Error,
+		"status_code", msg.StatusCode,
+	)
+}
