+++ coordinator/api/provider_register.go
@@ -1,0 +1,34 @@
+package api
+
+import (
+	"net/http"
+
+	"github.com/eigeninference/d-inference/coordinator/registry"
+)
+
+// attachProviderLocation resolves the provider's approximate geographic
+// location from the registration HTTP request. The resolved location is
+// stored on the Provider struct for stats aggregation. Raw IP addresses
+// are never persisted.
+func (s *Server) attachProviderLocation(providerID string, provider *registry.Provider, r *http.Request) {
+	if s.geoResolver == nil || provider == nil || r == nil {
+		return
+	}
+	loc := s.geoResolver.Lookup(r)
+	if loc == nil {
+		return
+	}
+	provider.Mu().Lock()
+	provider.Location = loc
+	provider.Mu().Unlock()
+	s.registry.PersistProvider(provider)
+	if s.readCache != nil {
+		s.readCache.Invalidate("stats:v1")
+	}
+	s.logger.Info("provider location resolved",
+		"provider_id", providerID,
+		"city", loc.City,
+		"country", loc.CountryCode,
+		"source", loc.Source,
+	)
+}
